Fix a problem when a request on port could be accepted before a web server is started up
Karma initially binds a port via setting up TCP net.Server to check whether the port is available.
And only then http.Server will be listened on top of TCP server.
It is possible that someone sends a request after port binding but before the web server is set up. In this case we need to propagate all collected sockets to the HTTP server